+++ relayer-cli/src/commands/query.rs
@@ -3,6 +3,7 @@
 use abscissa_core::{Command, Options, Runnable};
 
 use crate::commands::query::channels::QueryChannelsCmd;
+use crate::commands::query::packet::QueryPacketCmds;
 
 mod channel;
 mod channels;
@@ -83,31 +84,4 @@ pub enum QueryChannelCmds {
     /// The `query channel end` subcommand
     #[options(help = "Query channel end")]
     End(channel::QueryChannelEndCmd),
-}
-
-#[derive(Command, Debug, Options, Runnable)]
-pub enum QueryPacketCmds {
-    /// The `query packet commitments` subcommand
-    #[options(help = "Query packet commitments")]
-    Commitments(packet::QueryPacketCommitmentsCmd),
-
-    /// The `query packet commitment` subcommand
-    #[options(help = "Query packet commitment")]
-    Commitment(packet::QueryPacketCommitmentCmd),
-
-    /// The `query packet acks` subcommand
-    #[options(help = "Query packet acknowledgments")]
-    Acks(packet::QueryPacketAcknowledgementsCmd),
-
-    /// The `query packet ack` subcommand
-    #[options(help = "Query packet acknowledgment")]
-    Ack(packet::QueryPacketAcknowledgmentCmd),
-
-    /// The `query packet unreceived-packets` subcommand
-    #[options(help = "Query unreceived packets")]
-    UnreceivedPackets(packet::QueryUnreceivedPacketsCmd),
-
-    /// The `query packet unreceived-acks` subcommand
-    #[options(help = "Query unreceived acknowledgments")]
-    UnreceivedAcks(packet::QueryUnreceivedAcknowledgementCmd),
 }
